requirements.txt
psutil>=5.9.0
py-cpuinfo>=9.0.0
gputil>=1.4.0
speedtest-cli>=2.1.3
screeninfo>=0.8.1
requests>=2.28.0
wmi>=1.5.1; sys_platform == "win32"